+++ invoke_func.c
@@ -64,14 +64,14 @@ int invoke_fun(void * p) {
     return f(0, args);
 }
 
-long microseconds() {
+unsigned long microseconds() {
     struct timeval time;
     gettimeofday(&time, NULL);
     long microsec = ((unsigned long long)time.tv_sec * 1000000) + time.tv_usec;
     return microsec;
 }
 
-void microsleep(uint microseconds) {
+void microsleep(unsigned long microseconds) {
     struct timespec tim, tim2;
     tim.tv_sec = 0;
     tim.tv_nsec = microseconds * 1000;
